Reproducible Programming in R course -

RPiR is a learnr implementation of the Reproducible Programming in R course, developed by the Institute of Biodiversity, Animal Health, and Comparative Medicine at the University of Glasgow.

Details

To run a practical, you first need to load up the learnr package (see Examples). To reset a practical, click on the Start Over link (look for the tiny grey text below the left menu). To close a practical, close the window, then click on the red STOP button in RStudio to end the process.

The following tutorials are included in this package:

1-1 Exponential growth model; Writing scripts and commenting code

1-2 Exponential growth model; Writing functions

1-3 Exponential growth model; Sourcing code

1-4 Exponential growth model; Generalising and structuring code

1-5 Births and deaths model; Generalising code

1-6 Births and deaths model; Writing reports in RStudio

2-1 Creating an RStudio project

2-2 SIS model; Adapting previous code

2-3 SIS model; Changing the timestep

2-4 SIS model; Using while() loops

2-5 SIR model; Adapting previous code

3-1 Writing an R package

3-2 Stochastic population growth model; Using stop()

3-3 Stochastic births and deaths model; Passing functions as arguments

3-4 Stochastic SIS model

3-5 Stochastic SIR model

3-6 Compare the mean stochastic model to the deterministic SIR model

4-1 Biodiversity; Writing a data package

4-2 Biodiversity; Writing a package demo

4-3 Biodiversity; Adding package functions and documentation

For more information go to our GitHub page; https://github.com/IBAHCM/RPiR

Please raise an issue if you find any problems; https://github.com/IBAHCM/RPiR/issues

Author

Sonia Mitchell <sonia.mitchell@glasgow.ac.uk> (maintainer)
Richard Reeve <richard.reeve@glasgow.ac.uk>

Examples

if (FALSE) {
# To run a tutorial
library(learnr)
run_tutorial("practical1-1", package = "RPiR")
}