dummy_project

This function will create an RStudio project and populate it with scripts from practical series 1.

Usage

dummy_project(project_name)

Arguments

project_name

a string corresponding to the name you want to give your project; this should be your GitHub user name, followed by the practical number, *e.g.* maryjane_0106